Corrected multiple bugs, removed dependency on SteamManager, enabled support for Client Server as well as P2P networking
** VERY IMPORTANT **
Do not include Steamworks.NET in your build, that will cause issues when they use teh actual release Steamworks.NET as your modification is not inline with theres. It also means they cant use there own variations of the manager or any other 3rd party such as Heathen's Steamworks

- Removed FizzySteamworks.Awake
This was creating a steam_appid.txt if it was missing and that is a huge issue, that file should not be there in release so that was a major flaw. it also modified the app ID so if you typed in a different value here than you had in your steam_appid in your Editor and or you where using Heathen's Steamworks you would cause an issue where the user thought they where initalizing as 1 app but where actually running as another.

- Removed FetchSteamID
This would only work on clients and serves no real funciton other than initalizing relay which now has its own method and is platform sinsative.

- Added InitRelayNetworkAccess
This only funcitons in the case that UseNextGenSteamNetworking is true, it is platform sinsative e.g. uses GameServer API where needed

- Removed all uses of SteamManager.XXXX
SteamManager is an example class and has many many issues. It was only used to check for initalization of the API however every Steam API call already does this via TestIfAvailableClient and TestIfAvailableGameServer as required so we can simply remove all the if(SteamManager.Initalized) and instead use classic Try/Catch ... this also means this new version will work with Steamworks.NET SteamManager and Heathen's SteamSettings as well as any custom manager a user might create

- Removed unnessisary using statements

- Wrapped all API calls with platform check
#if UNITY_SERVER
	// Server only API calls
#else
	// Client only API calls
#endif

- NextCommon.ProcessMessage(IntPrt ptrs)
This has been made NotImplemented as it is not possible to support with the current release version of Steamworks.NET. If the user has implamented Steamworks.NET change set 424 then your previous code would throw an error. I have provided commented out code that will work with 424 so the user can choose to implament this if they follow the linked change set and or are using Heathen's Steamworks Foundation or Steamworks Complete which already has 424 implamented.
This is a temporary measure until Steamworks.NET releases an offical fix for the release issue which is in testing now

- Updated the README
Updated the notes on how to install
Removed the note about 480 ... 480 is the test app you are supposed to use it for testing its not that it isn't allowed
Updated the connection notes to support both P2P and Client Server
Added links to Heathen's Steamworks Foundation (free) and Steamworks Complete as both would now be supported by this transport as well as Steamworks.NET